+++ CHANGELOG.md
@@ -1,8 +1,23 @@
 # Changelog
 All notable changes to this project will be documented in this file using the standards as defined at [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0).
 
+## Version 4.2.0 *(2023-05-04)*
+
+Added:
+- Editor Window, `AdaptersWindow.cs`, for Chartboost Mediation adapters handling. For capabilities, see below:
+  * Ad Adapters are now decoupled from the Chartboost Mediation UPM package.
+  * Ad Adapters can be fetched, selected, and upgraded through the new Adapters Window.
+  * `Adapters Window` can be accessed through the `Chartboost Mediation/Adapters` menu.
+  * Public C# API is available to automate tasks such as upgrading, adding, loading, and saving Ad Adapter selections.
+    *Adapters information is automatically fetched on startup and manually through the `Adapters Window`.
+
+Improvements:
+- Fixed standardized error reporting of `ChartboostMediationError` components in iOS.
+
+API Changes:
+- Marked `ChartboostMediationIntegrationChecker.cs` as deprecated. Please utilize `AdaptersWindow.cs` instead.
+
 ## Version 4.1.0 *(2023-03-30)*
- 
 Improvements:
 - Simplified AppId & AppSignature accessors in `ChartboostMediationSettings`.
 - Added TestMode call to `ChartboostMediation.cs`.
@@ -98,15 +113,15 @@ Improvements:
 - Helium impression events are now separate from partner network impression events.
 - Sequential waterfall ad loading for all ad formats.
 - Removal of UnitySendMessage callbacks for proper Native callbacks on Android & iOS
-- `HeliumBannerAd`
+- `HeliumBannerAd`   
   - `Show` removed. Banners are ready to show upon successful load.
   - `ReadyToShow` removed since it is no longer necessary.
-- `HeliumSDK`
+- `HeliumSDK`   
   - `DidShowBanner` callback removed since it is no longer necessary.
   - Added the following callbacks for when Helium determines an ad to be visible on screen.
     - `DidRecordImpressionInterstitial`
     - `DidRecordImpressionRewarded`
-    - `DidRecordImpressionBanner`
+    - `DidRecordImpressionBanner`   
 
 Bug Fixes:
 - Ad load requests to the native Helium SDKs are dispatched to a background thread.